requirements.txt
streamlit
pymongo
plotly
pandas
wordcloud
matplotlib
python-dotenv
